Moved interface under Contract namespace
